+++ electron/preload.cjs
@@ -12,6 +12,8 @@ contextBridge.exposeInMainWorld('nebulaAPI', {
   artistInfo: (platform, artistId) => ipcRenderer.invoke('nebula:artist-info', { platform, artistId }),
   artistSongs: (platform, artistId) => ipcRenderer.invoke('nebula:artist-songs', { platform, artistId }),
   artistAlbums: (platform, artistId) => ipcRenderer.invoke('nebula:artist-albums', { platform, artistId }),
+  searchSongs: (keyword, pageSize) => ipcRenderer.invoke('nebula:search-songs', { keyword, pageSize }),
+  searchArtists: (keyword, pageSize) => ipcRenderer.invoke('nebula:search-artists', { keyword, pageSize }),
   loginQr: () => ipcRenderer.invoke('nebula:login:qr'),
   loginPoll: (unikey) => ipcRenderer.invoke('nebula:login:poll', { unikey }),
   loginPlatforms: () => ipcRenderer.invoke('nebula:login:platforms'),